@@ -1,10 +1,10 @@
 package io.github.jhipster.sample;
 
+import jakarta.annotation.Generated;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
-import javax.annotation.Generated;
 
 @Generated(value = "JHipster", comments = "Generated by JHipster 7.9.4")
 @Retention(RetentionPolicy.SOURCE)
