@@ -1,4 +1,4 @@
 /**
- * Data Transfer Objects.
+ * Data transfer objects for rest mapping.
  */
 package io.github.jhipster.sample.service.dto;
