@@ -1,14 +1,24 @@
 package io.github.jhipster.sample.web.rest.errors;
 
-import org.zalando.problem.AbstractThrowableProblem;
-import org.zalando.problem.Status;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.ErrorResponseException;
+import tech.jhipster.web.rest.errors.ProblemDetailWithCause.ProblemDetailWithCauseBuilder;
 
 @SuppressWarnings("java:S110") // Inheritance tree of classes should not be too deep
-public class InvalidPasswordException extends AbstractThrowableProblem {
+public class InvalidPasswordException extends ErrorResponseException {
 
     private static final long serialVersionUID = 1L;
 
     public InvalidPasswordException() {
-        super(ErrorConstants.INVALID_PASSWORD_TYPE, "Incorrect password", Status.BAD_REQUEST);
+        super(
+            HttpStatus.BAD_REQUEST,
+            ProblemDetailWithCauseBuilder
+                .instance()
+                .withStatus(HttpStatus.BAD_REQUEST.value())
+                .withType(ErrorConstants.INVALID_PASSWORD_TYPE)
+                .withTitle("Incorrect password")
+                .build(),
+            null
+        );
     }
 }
