@@ -15,7 +15,7 @@ describe('BankAccount e2e test', () => {
   const bankAccountPageUrlPattern = new RegExp('/bank-account(\\?.*)?$');
   const username = Cypress.env('E2E_USERNAME') ?? 'user';
   const password = Cypress.env('E2E_PASSWORD') ?? 'user';
-  const bankAccountSample = { name: 'Refined Visionary', balance: 14517 };
+  const bankAccountSample = { name: 'stamp gah', balance: 31518.71 };
 
   let bankAccount;
 
@@ -92,7 +92,7 @@ describe('BankAccount e2e test', () => {
             {
               statusCode: 200,
               body: [bankAccount],
-            }
+            },
           ).as('entitiesRequestInternal');
         });
 
@@ -157,9 +157,11 @@ describe('BankAccount e2e test', () => {
     });
 
     it('should create an instance of BankAccount', () => {
-      cy.get(`[data-cy="name"]`).type('experiences withdrawal').should('have.value', 'experiences withdrawal');
+      cy.get(`[data-cy="name"]`).type('quintessential lightly excited');
+      cy.get(`[data-cy="name"]`).should('have.value', 'quintessential lightly excited');
 
-      cy.get(`[data-cy="balance"]`).type('93119').should('have.value', '93119');
+      cy.get(`[data-cy="balance"]`).type('24759.29');
+      cy.get(`[data-cy="balance"]`).should('have.value', '24759.29');
 
       cy.get(entityCreateSaveButtonSelector).click();
 
