@@ -12,7 +12,6 @@
 // the project's config changing)
 import { existsSync, mkdirSync, writeFileSync } from 'fs';
 import { lighthouse, pa11y, prepareAudit } from 'cypress-audit';
-import ReportGenerator from 'lighthouse/report/generator/report-generator';
 
 export default async (on: Cypress.PluginEvents, config: Cypress.PluginConfigOptions) => {
   on('before:browser:launch', (browser, launchOptions) => {
@@ -23,9 +22,22 @@ export default async (on: Cypress.PluginEvents, config: Cypress.PluginConfigOpti
     }
   });
 
+  // Allows logging with cy.task('log', 'message') or cy.task('table', object)
   on('task', {
-    lighthouse: lighthouse(lighthouseReport => {
-      !existsSync('target/cypress') && mkdirSync('target/cypress', { recursive: true });
+    log(message) {
+      console.log(message);
+      return null;
+    },
+    table(message) {
+      console.table(message);
+      return null;
+    },
+  });
+
+  on('task', {
+    lighthouse: lighthouse(async lighthouseReport => {
+      const { default: ReportGenerator } = await import('lighthouse/report/generator/report-generator');
+      !existsSync('target/cypress/') && mkdirSync('target/cypress/', { recursive: true });
       writeFileSync('target/cypress/lhreport.html', ReportGenerator.generateReport(lighthouseReport.lhr, 'html'));
     }),
     pa11y: pa11y(),
