@@ -74,16 +74,22 @@ export const configurationPageHeadingSelector = '[data-cy="configurationPageHead
 // End Specific Selector Attributes for Cypress
 // ***********************************************
 
-export const classInvalid = 'invalid';
-export const classValid = 'valid';
+export const classInvalid = 'ng-invalid';
+
+export const classValid = 'ng-valid';
+
 Cypress.Commands.add('authenticatedRequest', data => {
-  const bearerToken = sessionStorage.getItem(Cypress.env('jwtStorageName'));
-  return cy.request({
-    ...data,
-    auth: {
-      bearer: bearerToken,
-    },
-  });
+  const jwtToken = sessionStorage.getItem(Cypress.env('jwtStorageName'));
+  const bearerToken = jwtToken && JSON.parse(jwtToken);
+  if (bearerToken) {
+    return cy.request({
+      ...data,
+      auth: {
+        bearer: bearerToken,
+      },
+    });
+  }
+  return cy.request(data);
 });
 
 Cypress.Commands.add('login', (username: string, password: string) => {
@@ -107,7 +113,7 @@ Cypress.Commands.add('login', (username: string, password: string) => {
       validate() {
         cy.authenticatedRequest({ url: '/api/account' }).its('status').should('eq', 200);
       },
-    }
+    },
   );
 });
 
