@@ -1,7 +1,8 @@
 function setupProxy({ tls }) {
+  const serverResources = ['/api', '/services', '/management', '/v3/api-docs', '/h2-console', '/auth', '/health'];
   const conf = [
     {
-      context: ['/api', '/services', '/management', '/v3/api-docs', '/h2-console', '/auth', '/health'],
+      context: serverResources,
       target: `http${tls ? 's' : ''}://localhost:8080`,
       secure: false,
       changeOrigin: tls,
